fix(processors): revert show all but show small nodes
Turning on show_all is unnecessary, to get nice graphs just removing the
insignificant nodes filter is enough. Here we can keep small nodes because they
don't take much pixels on a graph - as opposed to text output - but can still
become important when zooming in using the interactive display.